Move Yosemite dependent code to extension file building against main target only
