+++ WooCommerce/Classes/Model/MarkOrderAsReadUseCase.swift
@@ -6,10 +6,6 @@ import Networking
 import NetworkingCore
 #endif
 
-#if canImport(Yosemite) && !NOTIFICATION_EXTENSION
-import Yosemite
-#endif
-
 struct MarkOrderAsReadUseCase {
     /// Possible error states.
     ///
@@ -18,56 +14,6 @@ struct MarkOrderAsReadUseCase {
         case unavailableNote
         case noNeedToMarkAsRead
     }
-
-#if canImport(Yosemite) && !NOTIFICATION_EXTENSION
-    /// Async method that marks the order note as read if it is the notification for the last order.
-    /// We do it in a way that first we syncronize notification to get the remote `Note`
-    /// and then we compare local `orderID` with the one from remote `Note`.
-    /// If they match we mark it as read.
-    /// Returns syncronized note if marking was successful and error if some error happened
-    @MainActor
-    static func markOrderNoteAsReadIfNeeded(stores: StoresManager, noteID: Int64, orderID: Int) async -> Result<Note, Error> {
-        let syncronizedNoteResult: Result<Note, Error> = await withCheckedContinuation { continuation in
-            let action = NotificationAction.synchronizeNotification(noteID: noteID) { syncronizedNote, error in
-                guard let syncronizedNote = syncronizedNote else {
-                    continuation.resume(returning: .failure(MarkOrderAsReadUseCase.Error.unavailableNote))
-                    return
-                }
-                continuation.resume(returning: .success(syncronizedNote))
-            }
-            stores.dispatch(action)
-        }
-
-        switch syncronizedNoteResult {
-        case .success(let syncronizedNote):
-            guard let syncronizedNoteOrderID = syncronizedNote.meta.identifier(forKey: .order),
-                  syncronizedNoteOrderID == orderID,
-                  syncronizedNote.read == false else {
-                return .failure(MarkOrderAsReadUseCase.Error.noNeedToMarkAsRead)
-            }
-
-            let updateNoteStatusResult: Result<Note, Error> = await withCheckedContinuation { continuation in
-                let syncAction = NotificationAction.updateReadStatus(noteID: noteID, read: true) { error in
-                    if let error {
-                        continuation.resume(returning: .failure(MarkOrderAsReadUseCase.Error.failure(error)))
-                    } else {
-                        continuation.resume(returning: .success(syncronizedNote))
-                    }
-                }
-                stores.dispatch(syncAction)
-            }
-
-            switch updateNoteStatusResult {
-            case .success(let note):
-                return .success(note)
-            case .failure(let error):
-                return .failure(error)
-            }
-        case .failure(let error):
-            return .failure(error)
-        }
-    }
-#endif
 
     /// Async method that marks the order note as read if it is the notification for the last order.
     /// We do it in a way that first we syncronize notification to get the remote `Note`
